Fixed the really bad feeling movement
I didn't realize the Unity Input manager has JoyStick deadzone built in so I was trying to manage that myself. Becuase of this I needed to read in Horizontal and Vertical movement with GetAxis rather then GetAxisRaw. Since GetAxis smooths the transition from (1 || -1) -> 0 it was creating a very bad movement smoothing effect.